# apis/bijux-canon-index/v1/pinned_openapi.json
{
  "components": {
    "schemas": {
...
      "ExecuteResponse": {
        "additionalProperties": false,
        "description": "Represents execute response.",
        "properties": {
          "correlation_id": {
            "description": "Correlation identifier attached to the execution lifecycle.",
            "title": "Correlation Id",
            "type": "string"
          },
          "execution_contract": {
            "description": "Execution contract used to evaluate the query.",
            "title": "Execution Contract",
            "type": "string"
          },
          "execution_contract_status": {
            "description": "Human-readable stability marker for the chosen contract.",
            "title": "Execution Contract Status",
            "type": "string"
          },
          "execution_id": {
            "description": "Stable execution record identifier written by the run store.",
            "title": "Execution Id",
            "type": "string"
          },
          "replayable": {
            "description": "Whether the execution result can be replayed faithfully.",
            "title": "Replayable",
            "type": "boolean"
          },
          "results": {
            "description": "Ordered vector identifiers returned by the execution request.",
            "items": {
              "type": "string"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "results",
          "correlation_id",
          "execution_contract",
          "execution_contract_status",
          "replayable",
          "execution_id"
        ],
        "title": "ExecuteResponse",
        "type": "object"
      },
...
      "ExecutionBudgetPayload": {
        "additionalProperties": false,
        "description": "Represents execution budget payload.",
        "properties": {
          "max_error": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Error"
          },
          "max_latency_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Latency Ms"
          },
          "max_memory_mb": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Memory Mb"
          }
        },
        "title": "ExecutionBudgetPayload",
        "type": "object"
      },
      "ExecutionContract": {
        "description": "Enumeration of execution contract.",
        "enum": [
          "deterministic",
          "non_deterministic"
        ],
        "title": "ExecutionContract",
        "type": "string"
      },
...
      "ExecutionMode": {
        "description": "Enumeration of execution mode.",
        "enum": [
          "strict",
          "bounded",
          "exploratory"
        ],
        "title": "ExecutionMode",
        "type": "string"
      },
      "ExecutionRequestPayload": {
        "additionalProperties": false,
        "description": "Represents execution request payload.",
        "properties": {
          "artifact_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Id"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          },
          "execution_budget": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExecutionBudgetPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "execution_contract": {
            "$ref": "#/components/schemas/ExecutionContract"
          },
          "execution_intent": {
            "$ref": "#/components/schemas/ExecutionIntent"
          },
          "execution_mode": {
            "$ref": "#/components/schemas/ExecutionMode",
            "default": "strict"
          },
          "nd_adaptive_k": {
            "default": false,
            "title": "Nd Adaptive K",
            "type": "boolean"
          },
          "nd_build_on_demand": {
            "default": false,
            "title": "Nd Build On Demand",
            "type": "boolean"
          },
          "nd_candidate_k": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nd Candidate K"
          },
          "nd_diversity_lambda": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nd Diversity Lambda"
          },
          "nd_ef_construction": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nd Ef Construction"
          },
          "nd_ef_search": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nd Ef Search"
          },
          "nd_incremental_index": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nd Incremental Index"
          },
          "nd_latency_budget_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nd Latency Budget Ms"
          },
          "nd_low_signal_margin": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nd Low Signal Margin"
          },
          "nd_low_signal_refuse": {
            "default": false,
            "title": "Nd Low Signal Refuse",
            "type": "boolean"
          },
          "nd_m": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nd M"
          },
          "nd_max_candidates": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nd Max Candidates"
          },
          "nd_max_ef_search": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nd Max Ef Search"
          },
          "nd_max_index_memory_mb": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nd Max Index Memory Mb"
          },
          "nd_normalize_query": {
            "default": false,
            "title": "Nd Normalize Query",
            "type": "boolean"
          },
          "nd_normalize_vectors": {
            "default": false,
            "title": "Nd Normalize Vectors",
            "type": "boolean"
          },
          "nd_outlier_threshold": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nd Outlier Threshold"
          },
          "nd_profile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nd Profile"
          },
          "nd_replay_strict": {
            "default": false,
            "title": "Nd Replay Strict",
            "type": "boolean"
          },
          "nd_space": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nd Space"
          },
          "nd_target_recall": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nd Target Recall"
          },
          "nd_two_stage": {
            "default": true,
            "title": "Nd Two Stage",
            "type": "boolean"
          },
          "nd_warmup_queries": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nd Warmup Queries"
          },
          "nd_witness_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nd Witness Mode"
          },
          "nd_witness_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nd Witness Rate"
          },
          "nd_witness_sample_k": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nd Witness Sample K"
          },
          "randomness_profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RandomnessProfilePayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Text"
          },
          "top_k": {
            "default": 5,
            "exclusiveMinimum": 0.0,
            "title": "Top K",
            "type": "integer"
          },
          "vector": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vector"
          },
          "vector_store": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vector Store"
          },
          "vector_store_options": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vector Store Options"
          },
          "vector_store_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vector Store Uri"
          }
        },
        "required": [
          "execution_contract",
          "execution_intent"
        ],
        "title": "ExecutionRequestPayload",
        "type": "object"
      },
      "ExplainRequest": {
        "additionalProperties": false,
        "description": "Represents explain request.",
        "properties": {
          "artifact_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Id"
          },
          "result_id": {
            "minLength": 1,
            "title": "Result Id",
            "type": "string"
          }
        },
        "required": [
          "result_id"
        ],
        "title": "ExplainRequest",
        "type": "object"
      },
      "ExplainResponse": {
        "additionalProperties": false,
        "description": "Represents explain response.",
        "properties": {
          "artifact_id": {
            "description": "Execution artifact used to resolve the result.",
            "title": "Artifact Id",
            "type": "string"
          },
          "chunk_id": {
            "description": "Chunk that contributed the requested result.",
            "title": "Chunk Id",
            "type": "string"
          },
          "correlation_id": {
            "description": "Correlation identifier of the execution that produced the result.",
            "title": "Correlation Id",
            "type": "string"
          },
          "document_id": {
            "description": "Document that produced the requested result.",
            "title": "Document Id",
            "type": "string"
          },
          "execution_contract": {
            "description": "Execution contract used by the originating execution.",
            "title": "Execution Contract",
            "type": "string"
          },
          "execution_contract_status": {
            "description": "Human-readable stability marker for the chosen contract.",
            "title": "Execution Contract Status",
            "type": "string"
          },
          "execution_id": {
            "description": "Execution record identifier associated with the explained result.",
            "title": "Execution Id",
            "type": "string"
          },
          "metric": {
            "description": "Similarity metric used by the artifact.",
            "title": "Metric",
            "type": "string"
          },
          "replayable": {
            "description": "Whether the explained result belongs to a replayable artifact.",
            "title": "Replayable",
            "type": "boolean"
          },
          "score": {
            "description": "Score assigned to the result.",
            "title": "Score",
            "type": "number"
          },
          "vector_id": {
            "description": "Vector identifier for the explained result.",
            "title": "Vector Id",
            "type": "string"
          }
        },
        "required": [
          "document_id",
          "chunk_id",
          "vector_id",
          "artifact_id",
          "metric",
          "score",
          "correlation_id",
          "execution_contract",
          "execution_contract_status",
          "replayable",
          "execution_id"
        ],
        "title": "ExplainResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {
      "name": "Bijux",
      "url": "https://github.com/bijux/bijux-canon"
    },
    "description": "The bijux-canon-index HTTP API exposes the v1 surface for deterministic and bounded vector execution. It separates capability discovery, artifact materialization, and query execution so operators can review the contract they are invoking before they ship it.",
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "summary": "Contract-driven vector execution and replayable retrieval.",
    "title": "bijux-canon-index API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/execute": {
      "post": {
        "description": "Execute a deterministic or bounded retrieval request against a materialized artifact and return the result identifiers plus replay metadata.",
        "operationId": "executeVectorQuery",
        "parameters": [
          {
            "in": "header",
            "name": "X-Correlation-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "execute": {
                  "value": {
                    "artifact_id": "art-1",
                    "execution_contract": "deterministic",
                    "execution_intent": "exact_validation",
                    "execution_mode": "strict",
                    "top_k": 3,
                    "vector": [
                      0.0,
                      1.0,
                      0.0
                    ]
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/ExecutionRequestPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "execute": {
                    "value": {
                      "correlation_id": "req-example",
                      "execution_contract": "deterministic",
                      "execution_contract_status": "stable",
                      "execution_id": "exec-1",
                      "replayable": true,
                      "results": [
                        "vec-1"
                      ]
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ExecuteResponse"
                }
              }
            },
            "description": "Execution completed successfully."
          },
          "422": {
            "content": {
              "application/json": {
                "examples": {
                  "refusal": {
                    "value": {
                      "error": {
                        "message": "[INV-000] Deterministic execution requires a deterministic vector store",
                        "reason": "determinism_violation",
                        "remediation": "Use deterministic inputs or switch to non_deterministic contract with declared randomness."
                      }
                    }
                  }
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Execute a vector query",
        "tags": [
          "Execution"
        ]
      }
    },
    "/explain": {
      "post": {
        "description": "Resolve a result identifier back to the document, chunk, vector, and execution context that produced it.",
        "operationId": "explainExecutionResult",
        "parameters": [
          {
            "in": "header",
            "name": "X-Correlation-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "explain": {
                  "value": {
                    "result_id": "res-1"
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/ExplainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "explain": {
                    "value": {
                      "artifact_id": "art-1",
                      "chunk_id": "chunk-1",
                      "correlation_id": "req-example",
                      "document_id": "doc-1",
                      "execution_contract": "deterministic",
                      "execution_contract_status": "stable",
                      "execution_id": "exec-1",
                      "metric": "l2",
                      "replayable": true,
                      "score": 0.99,
                      "vector_id": "res-1"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ExplainResponse"
                }
              }
            },
            "description": "Result explanation returned successfully."
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "refusal": {
                    "value": {
                      "error": {
                        "message": "[INV-000] Deterministic execution requires a deterministic vector store",
                        "reason": "determinism_violation",
                        "remediation": "Use deterministic inputs or switch to non_deterministic contract with declared randomness."
                      }
                    }
                  }
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Explain a result",
        "tags": [
          "Execution"
        ]
      }
    },
...
    "/replay": {
      "post": {
        "description": "Replay a prior execution contract against the currently available artifact and report whether the replay stayed within the declared equivalence boundary.",
        "operationId": "replayExecutionOutcome",
        "parameters": [
          {
            "in": "header",
            "name": "X-Correlation-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "replay": {
                  "value": {
                    "artifact_id": "art-1",
                    "request_text": "hello"
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/ExecutionRequestPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "replay": {
                    "value": {
                      "details": {
                        "status": "matched"
                      },
                      "execution_contract": "deterministic",
                      "execution_contract_status": "stable",
                      "execution_id": "exec-1",
                      "matches": true,
                      "nondeterministic_sources": [],
                      "original_fingerprint": "orig-fp",
                      "replay_fingerprint": "replay-fp",
                      "replayable": true
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ReplayResponse"
                }
              }
            },
            "description": "Replay completed successfully."
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "refusal": {
                    "value": {
                      "error": {
                        "message": "[INV-000] Deterministic execution requires a deterministic vector store",
                        "reason": "determinism_violation",
                        "remediation": "Use deterministic inputs or switch to non_deterministic contract with declared randomness."
                      }
                    }
                  }
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replay an execution outcome",
        "tags": [
          "Execution"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
}